refactor: Moved file handling logic outside of signal handler to ensure handler is small and quick without any blocking functions

+++ animate_client.c
@@ -30,8 +30,8 @@ int main(int argc, char** argv, char** envp) {
     char fifo_c2s[64];
     char fifo_s2c[64];
 
-    sprintf(fifo_c2s, "FIFO_C2S_%d", getpid());
-    sprintf(fifo_s2c, "FIFO_S2C_%d", getpid());
+    sprintf(fifo_c2s, "FIFO_C2S_%ld", (long) getpid());
+    sprintf(fifo_s2c, "FIFO_S2C_%ld", (long) getpid());
 
     int fd_c2s = open(fifo_c2s, O_WRONLY);
     int fd_s2c = open(fifo_s2c, O_RDONLY);
